+++ jdaviz/configs/mosviz/tests/test_data_loading.py
@@ -157,6 +157,8 @@ def test_load_single_image_multi_spec(mosviz_helper, mos_image, spectrum1d, mos_
     spectra2d = [mos_spectrum2d] * 3
 
     image_viewer = mosviz_helper.app.get_viewer('image-viewer')
+    spec1d_viewer = mosviz_helper.app.get_viewer('spectrum-viewer')
+    spec2d_viewer = mosviz_helper.app.get_viewer('spectrum-2d-viewer')
 
     # Coordinates info panel should not crash even when nothing is loaded.
     image_viewer.on_mouse_or_key_event({'event': 'mouseover'})
@@ -180,14 +182,15 @@ def test_load_single_image_multi_spec(mosviz_helper, mos_image, spectrum1d, mos_
     assert len(qtable) == 3
 
     # Also check coordinates info panels for Mosviz image viewer.
-    # 1D spectrum viewer panel is already tested in Specviz.
-    # 2D spectrum viewer panel is already tested in Specviz2d.
+    # 1D spectrum viewer panel is also tested in Specviz.
+    # 2D spectrum viewer panel is also tested in Specviz2d.
 
     image_viewer.on_mouse_or_key_event({'event': 'mousemove', 'domain': {'x': 0, 'y': 0}})
     assert image_viewer.label_mouseover.pixel == 'x=000.0 y=000.0'
     assert image_viewer.label_mouseover.value == '+3.74540e-01 Jy'
     assert image_viewer.label_mouseover.world_ra_deg == '5.0297844783'
     assert image_viewer.label_mouseover.world_dec_deg == '4.9918991917'
+    assert image_viewer.label_mouseover.icon == 'a'
 
     image_viewer.on_mouse_or_key_event({'event': 'mousemove', 'domain': {'x': None, 'y': 0}})
     assert image_viewer.label_mouseover.pixel == ''
@@ -206,6 +209,21 @@ def test_load_single_image_multi_spec(mosviz_helper, mos_image, spectrum1d, mos_
     assert image_viewer.label_mouseover.value == ''
     assert image_viewer.label_mouseover.world_ra_deg == ''
     assert image_viewer.label_mouseover.world_dec_deg == ''
+
+    spec2d_viewer.on_mouse_or_key_event({'event': 'mousemove', 'domain': {'x': 10, 'y': 100}})
+    assert spec2d_viewer.label_mouseover.pixel == 'x=00010.0 y=00100.0'
+    assert spec2d_viewer.label_mouseover.value == '+8.12986e-01 '
+    assert spec2d_viewer.label_mouseover.world_ra_deg == ''
+    assert spec2d_viewer.label_mouseover.world_dec_deg == ''
+    assert spec2d_viewer.label_mouseover.icon == 'b'
+
+    spec1d_viewer.on_mouse_or_key_event({'event': 'mousemove', 'domain': {'x': 7000, 'y': 170}})
+    assert spec1d_viewer.label_mouseover.pixel == '7.00000e+03, 1.70000e+02'
+    assert spec1d_viewer.label_mouseover.world_label_prefix == 'Wave'
+    assert spec1d_viewer.label_mouseover.world_ra == '6.88889e+03 Angstrom (4 pix)'
+    assert spec1d_viewer.label_mouseover.world_label_prefix_2 == 'Flux'
+    assert spec1d_viewer.label_mouseover.world_ra_deg == '1.35436e+01 Jy'
+    assert spec1d_viewer.label_mouseover.icon == 'c'
 
 
 def test_zip_error(mosviz_helper, tmp_path):